Fixing MSVC warnings in snp_store.h.
git-svn-id: http://svn.seqan.de/seqan/trunk@11126 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/core/apps/snp_store/snp_store.h
@@ -1354,7 +1354,7 @@ getHomoProbs(THomoTable & cnks,
 #endif		
 		for(unsigned j = 0; j < length(qualitiesForward[i]); ++j)
 		{
-			qual = ordValue(qualitiesForward[i][j])-33;
+			qual = static_cast<double>(ordValue(qualitiesForward[i][j])-33);
 			//qual = rescale into regular log
 			if(j>=256) fk = fks[255];
 			else fk = fks[j];
@@ -1374,7 +1374,7 @@ getHomoProbs(THomoTable & cnks,
 #endif
 		for(unsigned j = 0; j < length(qualitiesReverse[i]); ++j)
 		{
-			qual = ordValue(qualitiesReverse[i][j])-33;
+			qual = static_cast<double>(ordValue(qualitiesReverse[i][j])-33);
 			if(j>=256) fk = fks[255];
 			else fk = fks[j];
 			sumE[i] += fk * qual;
